requirements.txt
blinker
colorama
fastapi
flake8
httpx
isort
mongoengine
mysql-connector-python
sentry-sdk
sqlalchemy
pymysql
pytest
python-dotenv
uvicorn
starlette
pillow
pydantic
pymongo
pyjwt
pytest-mock
python-dateutil
python-multipart
